labs/module-01/demos/demo-05-openai-via-open-ai-sdk/pyproject.toml
[project]
name = "openai-via-open-ai-sdk"
version = "0.1.0"
description = "Accessing OpenAI via OpenAI SDK demo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.0",
    "openai>=2.10.0"
]
